requirements.txt
streamlit
langchain
langchain-openai
chromadb
python-dotenv
tiktoken
langchain-text-splitters
